src/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
langchain==0.0.335
langchain-community==0.0.16
langchain-core==0.1.18
xgboost==2.0.2
scikit-learn==1.3.2
pandas==2.1.3
joblib==1.3.2
python-dotenv==1.0.0
groq==0.4.1
google-generativeai==0.3.2
supabase==2.3.0
PyPDF2==3.0.1
python-docx==1.0.1
chromadb==0.4.22
sentence-transformers==2.2.2
numpy==1.24.3 